ABCdeZ.jl

ABCdeZ.jl

Inference by abcdesmc!

Inference by abcdemc!

Distributions and Priors

Index

Inference by abcdesmc!

ABCdeZ.abcdesmc! — Function

bar(x[, y])

Compute the Bar index between x and y.

If y is unspecified, compute the Bar index between all pairs of columns of x.

Examples

julia> bar([1, 2], [1, 2])
1

Inference by abcdemc!

ABCdeZ.abcdemc! — Function

bar(x[, y])

Compute the Bar index between x and y.

If y is unspecified, compute the Bar index between all pairs of columns of x.

Examples

julia> bar([1, 2], [1, 2])
1

Distributions and Priors

ABCdeZ.Factored — Type

Factored{N} <: Distribution{Multivariate, MixedSupport}

A Distribution type that can be used to combine multiple UnivariateDistribution's (independently).

Examples

julia> prior = Factored(Normal(0, 1), Uniform(-1, 1))
Factored{2}(
p: (Normal{Float64}(μ=0.0, σ=1.0), Uniform{Float64}(a=-1.0, b=1.0))
)

Distributions.pdf — Function

pdf(d::Factored, x)

Function to evaluate the pdf of a Factored distribution object.

Distributions.logpdf — Function

logpdf(d::Factored, x)

Function to evaluate the logpdf of a Factored distribution object.

Base.rand — Function

rand(rng::AbstractRNG, factoreddist::Factored)

Function to sample one element from a Factored object.

Base.length — Function

length(p::Factored)

Returns the number of distributions contained in p.

Index

ABCdeZ.Factored

ABCdeZ.abcdemc!

ABCdeZ.abcdesmc!

Base.length

Base.rand

Distributions.logpdf

Distributions.pdf